+++ app/components/ui/chakra-ui/layout/index.ts
@@ -1,0 +1,1 @@
+export * from './CustomSidebar';
