--- app/lib/layout/Sidebar.tsx
@@ -1,42 +1,0 @@
-import { Box, useColorModeValue } from '~/components';
-import { NavLink } from '@remix-run/react';
-
-const Sidebar = () => {
-  const bg = useColorModeValue('gray.200', 'gray.600');
-  const bgActive = useColorModeValue('gray.100', 'gray.800');
-  const bgHover = useColorModeValue('gray.300', 'gray.700');
-
-  type NavLink = {
-    name: string;
-    to: string;
-  };
-  const links: NavLink[] = [{ name: 'Builder', to: '/dashboard/builder' }];
-
-  return (
-    <Box width={200} as="aside" paddingY={22} bg={bg}>
-      <Box as="nav" aria-label="Main navigation">
-        <Box as="ul" listStyleType="none" margin={0} padding={0}>
-          {links.map((link, i) => (
-            <Box as="li" key={i}>
-              <NavLink key={i} to={link.to}>
-                {({ isActive }) => (
-                  <Box
-                    paddingX={8}
-                    paddingY={2}
-                    bg={isActive ? bgActive : undefined}
-                    _hover={{ bg: bgHover }}
-                    transition="background-color 0.2s ease-in-out"
-                  >
-                    {link.name}
-                  </Box>
-                )}
-              </NavLink>
-            </Box>
-          ))}
-        </Box>
-      </Box>
-    </Box>
-  );
-};
-
-export default Sidebar;
