+++ app/routes/dashboard+/builder+/design-attributes+/container+/_container.tsx
@@ -1,11 +1,4 @@
-import {
-  Box,
-  Text,
-  Flex,
-  useColorModeValue,
-  Button,
-  Stack,
-} from '~/components';
+import { Box, Text, Flex, Button, Stack, CustomSidebar } from '~/components';
 import { DataFunctionArgs, json } from '@remix-run/node';
 import { NavLink, Outlet, useLoaderData } from '@remix-run/react';
 import { prisma } from '~/utils/db.server';
@@ -54,49 +47,6 @@ export default function ContainerPage() {
     );
   };
 
-  const ContainerList = () => {
-    const bg = useColorModeValue('gray.200', 'gray.600');
-    const bgHover = useColorModeValue('gray.300', 'gray.700');
-
-    return (
-      <Box
-        width={200}
-        as="aside"
-        paddingY={22}
-        bg={bg}
-        border="1px"
-        borderColor="gray.300"
-      >
-        <Box as="nav" aria-label="Main navigation">
-          {containers.length === 0 && (
-            <Box as="p" padding={2}>
-              No containers found.
-            </Box>
-          )}
-          <Box as="ul" listStyleType="none" margin={0} padding={0}>
-            {containers.map((container, i) => {
-              const { id, title } = container;
-              return (
-                <Box as="li" key={i}>
-                  <NavLink to={id}>
-                    <Box
-                      paddingX={8}
-                      paddingY={2}
-                      _hover={{ bg: bgHover }}
-                      transition="background-color 0.2s ease-in-out"
-                    >
-                      {title}
-                    </Box>
-                  </NavLink>
-                </Box>
-              );
-            })}
-          </Box>
-        </Box>
-      </Box>
-    );
-  };
-
   const ContainerContent = () => {
     return (
       <Flex flex="1" width="full" paddingY={22} borderRadius={6}>
@@ -112,7 +62,7 @@ export default function ContainerPage() {
       <ContainerOverview />
       <Flex flex={1} border="1px" borderColor="gray.300">
         <Flex flex={1}>
-          <ContainerList />
+          <CustomSidebar items={containers} itemType="containers" />
           <ContainerContent />
         </Flex>
       </Flex>
